requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.0.1
idna==3.6
ipdb==0.13.13
ipython==8.21.0
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
tzdata==2024.1
urllib3==2.2.0
wcwidth==0.2.13
